requirements.txt
openai>=1.12.0
pyyaml>=6.0.0
tqdm>=4.66.0